refactor: update directories and imports (#35)
* chore: move files

* chore: minor changes

* chore: add layouts alias

* refactor: rename newsletter card

* chore: move project card into own folder

* chore: move to layouts

* refactor: create single export for svg's

* chore: create styled folder

* refactor: create single export for styled elements

* chore: create navbar folder

* chore: create footer folder
