+++ components/styled/index.js
@@ -1,0 +1,2 @@
+export { StyledLink } from './link';
+export { SuccessMessage, ErrorMessage } from './message';
