+++ components/svg/index.js
@@ -1,0 +1,4 @@
+export { Logo } from './logo';
+export { Chakra } from './chakraLogo';
+export { HeroVisual } from './heroVisual';
+export { MagicMirrorLogo } from './mmLogo';
