+++ pages/projects.js
@@ -1,10 +1,10 @@
 import { Heading, Text } from '@chakra-ui/layout';
-import Container from '../components/container';
-import ContentWrapper from '../components/contentWrapper';
+import Container from '@/layouts/container';
+import ContentWrapper from '@/layouts/contentWrapper';
 
 const Projects = () => {
   return (
-    <Container>
+    <Container title="Projects | Greg Ogun">
       <ContentWrapper>
         <Heading as="h1" variant="h2">
           Projects 🎲
